3-15 updated level segments, built a few segments
Today's Additions:
3-15-2022
PC -
Added 11 lights to the ship, added some mood to the scene
added new trails to help add some more details.

World -
Lowered the Directional Light 0.2, changed to baby blue

Scripts -
In general, I added notes in each script of what I want to added
Added a Time Counter, would like to make this or one like it the world controler, so we can add pause functionality

Level Segments -
Added a new folder which holds the Level Segments I am discussing in my wiki page
There are currently 2 segments which are based on a simple focus and spacing of the bricks for collection.
Follow speed is set to 35 as 20 is painfully slow to interact with.
Want to add speed increase based on the collection