Added some new art, scripts, and adjusted some of the player settings
Added rock geo and materials to add to the levels.
Adjusted teh Scoring Script to reflect a t0 brick collection count. Anticipate this to be where loop counts can occur
Tweaked the trails, and added particles to the PC. looking very tron currently, and i like that.
In my scoring script, I tried to set up a loop to run, but I am bad at loops, they are foreign to me.