created a few changes
4-7-22
confirmed a better game experiences with mroe top down cam, but not all is committed. I like the idea of using the different cameras.

changed the overall resolution to match a phone, and portrait format seems the most suitable.

Spinning Focus coin in UI

Restructured segments to represent camera placement

will need to consider more changes in this format, maybe....