timer + slider
I have a timer and slider now in place, I will plan on add a slider ot coin count, and see about using them in combo to create a simple arcade set up for keeping the loops going.
It will be between a swtich statement or an array, but I think either could work.
timer ends, a pop up supposed to come up, but it flickered for a while, but I did a change, and changed it back, and the flick is gone. no idea what happened.