cleaned up the timer
converted the timer into an Enum instead of a simple loop, allows for triggering timed events.

Time to start playing with variables in the spline editor and level segments.
time to make the coin strands into patterns which allow for more color choices

testing timings on bonus time and bunny time