List of basic updates from Weekend's work
Build_10222_00
Updated scoring system to include Purple and Blue Coins in the counter
Added 3 levels which the User will toggle through based on simple coin collections
	there is not much difference in each level at this point, but still works to show level progression
		in time, will add sky box changes & music changes, requiring a little more coding, but not needed this week
There is a Pause function, with a respective UnPause function, still learning the use of (something.method.Toggle(!something.method.Toggle), where the "!" is like a toggle.
added a save function, but not implemented completely yet. the point of this is to grab data from the game play experience to implement into the written
Still an error when game pauses after one level, this is from not resetting the correct data points on each level load, addressing, but not a full game breaking bug
Simple UI panels for Pause, Level Progression, and Quit
Game should cycle through 2-3 levels fairly cleanly
Still a bug in "Continue Level" in first 2 levels.
Cleaned up the project in big ways, clearing out a lot of the discovery dev used to understand this project and skills

Build_10222_01
Fixed a set of errors coming from a newer Unity from one of my teammates, syncing into my older version.